사진 업로드 편집 기능 추가 ( 더 수정해야 할 부분 있음 - 리눅스 폴더/파일 삭제 부분)

+++ src/main/java/com/example/chattingweb/api/service/ServerApiService.java
@@ -204,13 +204,18 @@ public class ServerApiService implements ServerApiServiceIf{
     @Override
     @Transactional
     public int deleteUploadPhoto(String postId) throws IOException {
+        int result = 1;
         //삭제할 파일 path 가져오기
         List<PhotoDto> photoDto = serverApiRepository.selectPhotoToDelete(Integer.parseInt(postId));
         System.out.println("photoDto ===>"+ photoDto);
 
+        if(photoDto.size() != 0){
+            result = serverApiRepository.deletePhotoTable(Integer.parseInt(postId));
+        }
+
         //FileUtil.deletePhoto(photoDto); // 폴더 및 파일 삭제
 
-        return serverApiRepository.deletePhotoTable(Integer.parseInt(postId));
+        return result;
     }
 
     @Override
